Move Presenting, Dumping and ImGui out of Renderer
