+++ Source/Core/VideoCommon/RenderBase.h
@@ -23,16 +23,10 @@
 #include <vector>
 
 #include "Common/CommonTypes.h"
-#include "Common/Event.h"
 #include "Common/Flag.h"
 #include "Common/MathUtil.h"
-#include "VideoCommon/AsyncShaderCompiler.h"
-#include "VideoCommon/BPMemory.h"
-#include "VideoCommon/FrameDump.h"
 #include "VideoCommon/GraphicsModSystem/Runtime/GraphicsModManager.h"
-#include "VideoCommon/PerformanceMetrics.h"
 #include "VideoCommon/RenderState.h"
-#include "VideoCommon/TextureConfig.h"
 
 class AbstractFramebuffer;
 class AbstractPipeline;
@@ -41,23 +35,22 @@ class AbstractTexture;
 class AbstractStagingTexture;
 class BoundingBox;
 class NativeVertexFormat;
-class NetPlayChatUI;
 class PixelShaderManager;
 class PointerWrap;
-struct TextureConfig;
 struct ComputePipelineConfig;
 struct AbstractPipelineConfig;
 struct PortableVertexDeclaration;
+struct TextureConfig;
+enum class AbstractTextureFormat : u32;
 enum class ShaderStage;
 enum class EFBAccessType;
 enum class EFBReinterpretType;
 enum class StagingTextureType;
-enum class AspectMode;
 
 namespace VideoCommon
 {
-class PostProcessing;
-}  // namespace VideoCommon
+class AsyncShaderCompiler;
+}
 
 struct EfbPokeData
 {
@@ -145,11 +138,6 @@ public:
   // Ideal internal resolution - multiple of the native EFB resolution
   int GetTargetWidth() const { return m_target_width; }
   int GetTargetHeight() const { return m_target_height; }
-  // Display resolution
-  int GetBackbufferWidth() const { return m_backbuffer_width; }
-  int GetBackbufferHeight() const { return m_backbuffer_height; }
-  float GetBackbufferScale() const { return m_backbuffer_scale; }
-  void SetWindowSize(int width, int height);
 
   // Sets viewport and scissor to the specified rectangle. rect is assumed to be in framebuffer
   // coordinates, i.e. lower-left origin in OpenGL.
@@ -174,25 +162,6 @@ public:
   // Use this to convert a whole native EFB rect to backbuffer coordinates
   MathUtil::Rectangle<int> ConvertEFBRectangle(const MathUtil::Rectangle<int>& rc) const;
 
-  const MathUtil::Rectangle<int>& GetTargetRectangle() const { return m_target_rectangle; }
-  float CalculateDrawAspectRatio() const;
-
-  // Crops the target rectangle to the framebuffer dimensions, reducing the size of the source
-  // rectangle if it is greater. Works even if the source and target rectangles don't have a
-  // 1:1 pixel mapping, scaling as appropriate.
-  void AdjustRectanglesToFitBounds(MathUtil::Rectangle<int>* target_rect,
-                                   MathUtil::Rectangle<int>* source_rect, int fb_width,
-                                   int fb_height);
-
-  std::tuple<float, float> ScaleToDisplayAspectRatio(int width, int height) const;
-  void UpdateDrawRectangle();
-
-  std::tuple<float, float> ApplyStandardAspectCrop(float width, float height) const;
-
-  // Use this to convert a single target rectangle to two stereo rectangles
-  std::tuple<MathUtil::Rectangle<int>, MathUtil::Rectangle<int>>
-  ConvertStereoRectangle(const MathUtil::Rectangle<int>& rc) const;
-
   unsigned int GetEFBScale() const;
 
   // Use this to upscale native EFB coordinates to IDEAL internal resolution
@@ -202,10 +171,6 @@ public:
   // Floating point versions of the above - only use them if really necessary
   float EFBToScaledXf(float x) const;
   float EFBToScaledYf(float y) const;
-
-  // Random utilities
-  void SaveScreenshot(std::string filename);
-  void DrawDebugText();
 
   virtual void ClearScreen(const MathUtil::Rectangle<int>& rc, bool colorEnable, bool alphaEnable,
                            bool zEnable, u32 color, u32 z);
@@ -230,12 +195,18 @@ public:
   void Swap(u32 xfb_addr, u32 fb_width, u32 fb_stride, u32 fb_height, u64 ticks);
 
   void UpdateWidescreenHeuristic();
+  bool IsGameWidescreen() const { return m_is_game_widescreen; }
 
-  // Draws the specified XFB buffer to the screen, performing any post-processing.
-  // Assumes that the backbuffer has already been bound and cleared.
-  virtual void RenderXFBToScreen(const MathUtil::Rectangle<int>& target_rc,
-                                 const AbstractTexture* source_texture,
-                                 const MathUtil::Rectangle<int>& source_rc);
+  // A simple presentation fallback, only used by video software
+  virtual void ShowImage(const AbstractTexture* source_texture,
+                         const MathUtil::Rectangle<int>& source_rc)
+  {
+  }
+
+  // For opengl's glDrawBuffer
+  virtual void SelectLeftBuffer() {}
+  virtual void SelectRightBuffer() {}
+  virtual void SelectMainBuffer() {}
 
   // Called when the configuration changes, and backend structures need to be updated.
   virtual void OnConfigChanged(u32 bits) {}
@@ -243,11 +214,7 @@ public:
   PixelFormat GetPrevPixelFormat() const { return m_prev_efb_format; }
   void StorePixelFormat(PixelFormat new_format) { m_prev_efb_format = new_format; }
   bool EFBHasAlphaChannel() const;
-  VideoCommon::PostProcessing* GetPostProcessor() const { return m_post_processor.get(); }
-  // Final surface changing
-  // This is called when the surface is resized (WX) or the window changes (Android).
-  void ChangeSurface(void* new_surface_handle);
-  void ResizeSurface();
+
   bool UseVertexDepthRange() const;
   void DoState(PointerWrap& p);
 
@@ -257,22 +224,11 @@ public:
   // interface and final XFB.
   bool UseGeometryShaderForUI() const;
 
-  // Returns a lock for the ImGui mutex, enabling data structures to be modified from outside.
-  // Use with care, only non-drawing functions should be called from outside the video thread,
-  // as the drawing is tied to a "frame".
-  std::unique_lock<std::mutex> GetImGuiLock();
-
-  // Begins/presents a "UI frame". UI frames do not draw any of the console XFB, but this could
-  // change in the future.
-  void BeginUIFrame();
-  void EndUIFrame();
-
   // Will forcibly reload all textures on the next swap
   void ForceReloadTextures();
 
   const GraphicsModManager& GetGraphicsModManager() const;
 
-protected:
   // Bitmask containing information about which configuration has changed for the backend.
   enum ConfigChangeBits : u32
   {
@@ -286,6 +242,7 @@ protected:
     CONFIG_CHANGE_BIT_BBOX = (1 << 7)
   };
 
+protected:
   std::tuple<int, int> CalculateTargetScale(int x, int y) const;
   bool CalculateTargetSize();
 
@@ -294,35 +251,10 @@ protected:
   void CheckFifoRecording();
   void RecordVideoMemory();
 
-  // ImGui initialization depends on being able to create textures and pipelines, so do it last.
-  bool InitializeImGui();
-
-  // Recompiles ImGui pipeline - call when stereo mode changes.
-  bool RecompileImGuiPipeline();
-
-  // Sets up ImGui state for the next frame.
-  // This function itself acquires the ImGui lock, so it should not be held.
-  void BeginImGuiFrame();
-
-  // Same as above but without locking the ImGui lock.
-  void BeginImGuiFrameUnlocked();
-
-  // Destroys all ImGui GPU resources, must do before shutdown.
-  void ShutdownImGui();
-
-  // Renders ImGui windows to the currently-bound framebuffer.
-  // Should be called with the ImGui lock held.
-  void DrawImGui();
-
   virtual std::unique_ptr<BoundingBox> CreateBoundingBox() const = 0;
 
   AbstractFramebuffer* m_current_framebuffer = nullptr;
   const AbstractPipeline* m_current_pipeline = nullptr;
-
-  Common::Flag m_screenshot_request;
-  Common::Event m_screenshot_completed;
-  std::mutex m_screenshot_lock;
-  std::string m_screenshot_name;
 
   bool m_is_game_widescreen = false;
   bool m_was_orthographically_anamorphic = false;
@@ -331,72 +263,12 @@ protected:
   int m_target_width = 1;
   int m_target_height = 1;
 
-  // Backbuffer (window) size and render area
-  int m_backbuffer_width = 0;
-  int m_backbuffer_height = 0;
-  float m_backbuffer_scale = 1.0f;
-  AbstractTextureFormat m_backbuffer_format = AbstractTextureFormat::Undefined;
-  MathUtil::Rectangle<int> m_target_rectangle = {};
   int m_frame_count = 0;
 
-  std::unique_ptr<VideoCommon::PostProcessing> m_post_processor;
-
-  void* m_new_surface_handle = nullptr;
-  Common::Flag m_surface_changed;
-  Common::Flag m_surface_resized;
-  std::mutex m_swap_mutex;
-
-  // ImGui resources.
-  std::unique_ptr<NativeVertexFormat> m_imgui_vertex_format;
-  std::vector<std::unique_ptr<AbstractTexture>> m_imgui_textures;
-  std::unique_ptr<AbstractPipeline> m_imgui_pipeline;
-  std::mutex m_imgui_mutex;
-  u64 m_imgui_last_frame_time;
-
 private:
-  std::tuple<int, int> CalculateOutputDimensions(int width, int height) const;
-
   PixelFormat m_prev_efb_format = PixelFormat::INVALID_FMT;
   unsigned int m_efb_scale = 1;
 
-  // These will be set on the first call to SetWindowSize.
-  int m_last_window_request_width = 0;
-  int m_last_window_request_height = 0;
-
-  // frame dumping:
-  FrameDump m_frame_dump;
-  std::thread m_frame_dump_thread;
-  Common::Flag m_frame_dump_thread_running;
-
-  // Used to kick frame dump thread.
-  Common::Event m_frame_dump_start;
-
-  // Set by frame dump thread on frame completion.
-  Common::Event m_frame_dump_done;
-
-  // Holds emulation state during the last swap when dumping.
-  FrameDump::FrameState m_last_frame_state;
-
-  // Communication of frame between video and dump threads.
-  FrameDump::FrameData m_frame_dump_data;
-
-  // Texture used for screenshot/frame dumping
-  std::unique_ptr<AbstractTexture> m_frame_dump_render_texture;
-  std::unique_ptr<AbstractFramebuffer> m_frame_dump_render_framebuffer;
-
-  // Double buffer:
-  std::unique_ptr<AbstractStagingTexture> m_frame_dump_readback_texture;
-  std::unique_ptr<AbstractStagingTexture> m_frame_dump_output_texture;
-  // Set when readback texture holds a frame that needs to be dumped.
-  bool m_frame_dump_needs_flush = false;
-  // Set when thread is processing output texture.
-  bool m_frame_dump_frame_running = false;
-
-  // Used to generate screenshot names.
-  u32 m_frame_dump_image_counter = 0;
-
-  // Tracking of XFB textures so we don't render duplicate frames.
-  u64 m_last_xfb_id = std::numeric_limits<u64>::max();
   u64 m_last_xfb_ticks = 0;
   u32 m_last_xfb_addr = 0;
   u32 m_last_xfb_width = 0;
@@ -416,40 +288,6 @@ private:
   // This produces much better results than just returning garbage, which can cause games like
   // Ultimate Spider-Man to crash
   std::array<u16, 4> m_bounding_box_fallback = {};
-
-  // NOTE: The methods below are called on the framedumping thread.
-  void FrameDumpThreadFunc();
-  bool StartFrameDumpToFFMPEG(const FrameDump::FrameData&);
-  void DumpFrameToFFMPEG(const FrameDump::FrameData&);
-  void StopFrameDumpToFFMPEG();
-  std::string GetFrameDumpNextImageFileName() const;
-  bool StartFrameDumpToImage(const FrameDump::FrameData&);
-  void DumpFrameToImage(const FrameDump::FrameData&);
-
-  void ShutdownFrameDumping();
-
-  bool IsFrameDumping() const;
-
-  // Checks that the frame dump render texture exists and is the correct size.
-  bool CheckFrameDumpRenderTexture(u32 target_width, u32 target_height);
-
-  // Checks that the frame dump readback texture exists and is the correct size.
-  bool CheckFrameDumpReadbackTexture(u32 target_width, u32 target_height);
-
-  // Fills the frame dump staging texture with the current XFB texture.
-  void DumpCurrentFrame(const AbstractTexture* src_texture,
-                        const MathUtil::Rectangle<int>& src_rect, u64 ticks, int frame_number);
-
-  // Asynchronously encodes the specified pointer of frame data to the frame dump.
-  void DumpFrameData(const u8* data, int w, int h, int stride);
-
-  // Ensures all rendered frames are queued for encoding.
-  void FlushFrameDump();
-
-  // Ensures all encoded frames have been written to the output file.
-  void FinishFrameData();
-
-  std::unique_ptr<NetPlayChatUI> m_netplay_chat_ui;
 
   Common::Flag m_force_reload_textures;
